refactor: call anySignal at the call sites instead of through a helper
The three call sites already only combine when there is a caller signal,
so a combineSignals wrapper duplicated that check and reordered the
arguments for nothing. Swap AbortSignal.any for anySignal in place, drop
the helper and the tests that only exercised it, and remove the two
comments that described WebKit's AbortSignal.any behavior, which is off
the live path now.
