+++ src/createRequester.ts
@@ -1,4 +1,5 @@
-import {combineSignals} from './combineSignals'
+import {anySignal} from 'any-signal'
+
 import {HttpError, TimeoutError} from './errors'
 import {createBufferedResponse} from './response'
 import type {
@@ -136,7 +137,7 @@ export function createRequester(
       // Reject via Promise.race rather than relying on fetch to reject with the
       // abort reason: workerd's fetch reconstructs the reason (losing its
       // prototype, so `instanceof TimeoutError` breaks), and WebKit has dropped
-      // the reason — or ignored aborts from a combined signal entirely.
+      // the reason.
       // In rejection-only mode there is no controller — the race alone rejects.
       controller = attachSignal ? new AbortController() : undefined
       deadlines.push(
@@ -151,7 +152,10 @@ export function createRequester(
 
     // Without a controller (rejection-only mode) the init is passed through
     // untouched, so a caller-provided signal reaches fetch as-is.
-    const signal = controller ? combineSignals(controller.signal, init.signal) : init.signal
+    const signal =
+      controller && init.signal
+        ? anySignal([init.signal, controller.signal])
+        : (controller?.signal ?? init.signal)
     // fetchFn is invoked inside the try so a synchronous throw still clears
     // the headers timer — otherwise the orphaned deadline would later reject
     // with nothing subscribed to it.
@@ -559,10 +563,6 @@ function buildFetchArgs(
   let signal: AbortSignal | undefined = opts.signal
   let clearTotalTimer: (() => void) | undefined
   if (totalMs !== undefined) {
-    // Own the deadline timer instead of using AbortSignal.timeout(): WebKit
-    // can garbage-collect an otherwise-unreferenced timeout signal behind a
-    // combined signal, silently disarming the deadline. The timer callback
-    // closure keeps this controller (and thus the abort chain) alive.
     const totalController = new AbortController()
     const timer = setTimeout(
       () =>
@@ -573,7 +573,7 @@ function buildFetchArgs(
     )
     unrefTimer(timer)
     clearTotalTimer = () => clearTimeout(timer)
-    signal = combineSignals(totalController.signal, signal)
+    signal = signal ? anySignal([signal, totalController.signal]) : totalController.signal
   }
   if (signal) init.signal = signal
